# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.AttackEventAggregateResult:
    description: 攻击事件的聚合统计结果，提供IP、域名、端口维度的攻击信息汇总，包含攻击次数、首次和最近攻击时间、持续时间等关键指标
    properties:
      count:
        description: 攻击总次数，同一来源的攻击计数
        example: 15
        type: integer
      domain:
        description: 域名，被攻击的站点
        example: example.com
        type: string
      dstPort:
        description: 目标端口号，被攻击的服务端口
        example: 443
        type: integer
      durationInMinutes:
        description: 攻击持续时间(分钟)，从首次到最近攻击的时间跨度
        example: 18.2
        type: number
      firstAttackTime:
        description: 首次攻击时间，该IP首次发起攻击的时间点
        example: "2024-03-18T08:12:33Z"
        type: string
      isOngoing:
        description: 是否正在进行中，标识攻击是否仍在持续
        example: true
        type: boolean
      lastAttackTime:
        description: 最近攻击时间，该IP最后一次攻击的时间点
        example: "2024-03-18T08:30:45Z"
        type: string
      srcIp:
        description: 来源IP地址，攻击者地址
        example: 192.168.1.100
        type: string
    type: object
  dto.AttackEventResponse:
    description: 攻击事件查询的分页响应结构体，包含聚合结果列表及分页元数据，用于前端展示和翻页控制
    properties:
      currentPage:
        description: 当前页码，从1开始计数
        example: 1
        type: integer
      pageSize:
        description: 每页大小，当前设置的每页记录数
        example: 10
        type: integer
      results:
        description: 聚合结果列表，当前页的攻击事件记录
        items:
          $ref: '#/definitions/dto.AttackEventAggregateResult'
        type: array
      totalCount:
        description: 总记录数，符合条件的攻击事件总数
        example: 35
        type: integer
      totalPages:
        description: 总页数，根据总记录数和每页大小计算
        example: 4
        type: integer
    type: object
  dto.AttackLogResponse:
    description: 攻击日志查询的分页响应结构体，返回详细的WAF日志记录及分页信息，便于安全分析和事件追踪
    properties:
      currentPage:
        description: 当前页码，从1开始计数
        example: 1
        type: integer
      pageSize:
        description: 每页大小，当前设置的每页记录数
        example: 10
        type: integer
      results:
        description: 日志记录列表，当前页的WAF攻击日志详情
        items:
          $ref: '#/definitions/model.WAFLog'
        type: array
      totalCount:
        description: 总记录数，符合查询条件的日志总数
        example: 128
        type: integer
      totalPages:
        description: 总页数，根据总记录数和每页大小计算
        example: 13
        type: integer
    type: object
...
  model.ErrResponse:
    description: 错误的API响应标准格式
    properties:
      code:
        example: 400
        type: integer
      error:
        example: 参数错误
        type: string
      message:
        example: 请求参数错误
        type: string
      requestId:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      success:
        example: false
        type: boolean
      timestamp:
        example: "2023-01-01T12:00:00Z"
        type: string
    type: object
  model.ErrResponseDontShowError:
    description: 错误的API响应标准格式,不展示 error
    properties:
      code:
        example: 400
        type: integer
      message:
        example: 请求参数错误
        type: string
      requestId:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      success:
        example: false
        type: boolean
      timestamp:
        example: "2023-01-01T12:00:00Z"
        type: string
    type: object
  model.Log:
    description: 详细的WAF规则匹配记录，包含规则触发的详细信息和原始日志
    properties:
      accuracy:
        description: 规则匹配准确度(0-10)
        example: 9
        type: integer
      logRaw:
        description: 原始日志数据
        example: '[2024-03-18 08:12:33] [error] ModSecurity: Access denied with code
          403 (phase 1). Matched "Operator ''Rx'' with parameter ''(?:scanner)'' against
          variable ''REQUEST_HEADERS:User-Agent''" [id "10086"] [msg "恶意扫描器检测"] [severity
          "CRITICAL"]'
        type: string
      message:
        description: 日志消息
        example: 恶意扫描器检测
        type: string
      payload:
        description: 攻击载荷
        example: Scanner/1.0
        type: string
      phase:
        description: 请求处理阶段
        example: 1
        type: integer
      ruleId:
        description: 规则ID
        example: 10086
        type: integer
      secLangRaw:
        description: 安全规则原始定义
        example: SecRule REQUEST_HEADERS:User-Agent "@rx (?:scanner)" "id:1008,phase:1,severity:'CRITICAL'"
        type: string
      secMark:
        description: 安全标记
        example: web_scanner
        type: string
      severity:
        description: 严重级别(0-5)
        example: 2
        type: integer
    type: object
...
  model.SuccessResponse:
    description: 成功的API响应标准格式
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        example: 操作成功
        type: string
      requestId:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      success:
        example: true
        type: boolean
      timestamp:
        example: "2023-01-01T12:00:00Z"
        type: string
    type: object
...
  model.WAFLog:
    description: Web应用防火墙安全事件完整记录，包含详细的攻击检测和防护信息
    properties:
      accuracy:
        description: 规则匹配准确度(0-10)
        example: 9
        type: integer
      createdAt:
        description: 事件发生时间戳
        example: "2024-03-18T08:12:33Z"
        type: string
      domain:
        description: 目标域名
        example: api.example.com
        type: string
      dstIp:
        description: 目标IP地址
        example: 10.0.0.1
        type: string
      dstPort:
        description: 目标端口
        example: 443
        type: integer
      id:
        description: 日志唯一标识符
        type: string
      logs:
        description: 关联的日志条目
        items:
          $ref: '#/definitions/model.Log'
        type: array
      message:
        description: 事件描述消息
        example: 恶意扫描器检测
        type: string
      payload:
        description: 攻击载荷
        example: Scanner/1.0
        type: string
      phase:
        description: 请求处理阶段
        example: 1
        type: integer
      request:
        description: 原始HTTP请求
        example: |-
          GET /api/v1/users HTTP/1.1
          Host: api.example.com
          User-Agent: Scanner/1.0
        type: string
      requestId:
        description: 请求唯一标识
        example: a1b2c3d4e5f6
        type: string
      response:
        description: 原始HTTP响应
        example: |-
          HTTP/1.1 403 Forbidden
          Content-Type: text/html
          Content-Length: 146
        type: string
      ruleId:
        description: 触发的规则ID
        example: 10086
        type: integer
      secLangRaw:
        description: 安全规则原始定义
        example: SecRule REQUEST_HEADERS:User-Agent "@rx (?:scanner)" "id:1008,phase:1,severity:'CRITICAL'"
        type: string
      secMark:
        description: 安全标记
        example: web_scanner
        type: string
      severity:
        description: 事件严重级别(0-5)
        example: 2
        type: integer
      srcIp:
        description: 来源IP地址
        example: 192.168.1.1
        type: string
      srcPort:
        description: 来源端口
        example: 52134
        type: integer
      uri:
        description: 请求URI路径
        example: /api/v1/users
        type: string
    type: object
...
host: localhost:2333
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 简单的 Web 应用防火墙管理系统 API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Simple-WAF API
  version: "1.0"
paths:
...
  /api/v1/waf/logs:
    get:
      consumes:
      - application/json
      description: 查询详细的WAF攻击日志记录，提供多条件筛选和分页功能，支持按规则ID、IP、域名、端口和时间范围过滤
      parameters:
      - description: 规则ID，触发攻击检测的WAF规则标识
        in: query
        name: ruleId
        type: integer
      - description: 来源IP地址，攻击者地址
        in: query
        name: srcIp
        type: string
      - description: 目标IP地址，被攻击的服务器地址
        in: query
        name: dstIp
        type: string
      - description: 域名，被攻击的站点域名
        in: query
        name: domain
        type: string
      - description: 来源端口号，发起攻击的端口
        in: query
        name: srcPort
        type: integer
      - description: 目标端口号，被攻击的服务端口
        in: query
        name: dstPort
        type: integer
      - description: 请求ID，唯一标识HTTP请求的ID
        in: query
        name: requestId
        type: string
      - description: '查询起始时间 (ISO8601格式，如: 2024-03-17T00:00:00Z)'
        in: query
        name: startTime
        type: string
      - description: '查询结束时间 (ISO8601格式，如: 2024-03-18T23:59:59Z)'
        in: query
        name: endTime
        type: string
      - description: '当前页码，从1开始计数 (默认: 1)'
        in: query
        name: page
        type: integer
      - description: '每页记录数，最大100条 (默认: 10)'
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.AttackLogResponse'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/model.ErrResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
      summary: 获取详细攻击日志
      tags:
      - WAF安全日志
  /api/v1/waf/logs/events:
    get:
      consumes:
      - application/json
      description: 按来源IP、目标端口和域名聚合的攻击事件统计，支持多维度筛选和分页
      parameters:
      - description: 来源IP地址，攻击者地址
        in: query
        name: srcIp
        type: string
      - description: 目标IP地址，被攻击的服务器地址
        in: query
        name: dstIp
        type: string
      - description: 域名，被攻击的站点域名
        in: query
        name: domain
        type: string
      - description: 来源端口号，发起攻击的端口
        in: query
        name: srcPort
        type: integer
      - description: 目标端口号，被攻击的服务端口
        in: query
        name: dstPort
        type: integer
      - description: '查询起始时间 (ISO8601格式，如: 2024-03-17T00:00:00Z)'
        in: query
        name: startTime
        type: string
      - description: '查询结束时间 (ISO8601格式，如: 2024-03-18T23:59:59Z)'
        in: query
        name: endTime
        type: string
      - description: '当前页码，从1开始计数 (默认: 1)'
        in: query
        name: page
        type: integer
      - description: '每页记录数，最大100条 (默认: 10)'
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.AttackEventResponse'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/model.ErrResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
      summary: 获取聚合攻击事件
      tags:
      - WAF安全日志
...
securityDefinitions:
  BearerAuth:
    description: 使用 Bearer {token} 格式进行身份验证
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"